XRDDEV-963 refactor
