+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/openapi/ClientsApiController.java
@@ -584,7 +584,7 @@ public class ClientsApiController implements ClientsApi {
             accessRights = accessRightConverter.convert(
                     serviceClientService.getServiceClientAccessRights(clientIdentifier, serviceClientId));
         } catch (IdentifierNotFoundException | ClientNotFoundException
-                | ServiceClientNotFoundException | LocalGroupNotFoundException e) {
+                | LocalGroupNotFoundException e) {
             throw new ResourceNotFoundException(e);
         }
         return new ResponseEntity<>(accessRights, HttpStatus.OK);
